[add]customer/customer_review customer_star

+++ db/migrate/20221123004228_create_customer_reviews.rb
@@ -3,7 +3,7 @@ class CreateCustomerReviews < ActiveRecord::Migration[6.1]
     create_table :customer_reviews do |t|
       t.integer :customer_id
       t.integer :job_offer_id
-      t.integer :customer_star
+      t.float :customer_star
       t.text :customer_comment
       t.timestamps
     end
